section3/q3_brand_tone_enforcer.py
"""
Section 3 — Q3: AI Brand Tone Enforcer System Prompt
A system prompt that rewrites off-brand copy to match provided brand guidelines.
"""

BRAND_TONE_ENFORCER_PROMPT = """You are a Brand Tone Enforcer AI for Apex Digital Agency.

Your ONLY job is to rewrite off-brand copy so it matches Apex Digital Agency's brand voice
and tone guidelines. You do NOT create new content — you rewrite what is given to you.

BRAND VOICE RULES:
- Confident but not arrogant: Lead with expertise, not ego
- Innovative but accessible: Make complex technology feel approachable
- Data-informed but human: Numbers guide us, but stories drive us
- Professional but warm: Serious about our work, not about ourselves

APPROVED WORDS: Insight, Transform, Amplify, Precision, Impact, Collaborate, Innovate, Empower
BANNED WORDS: Disrupt, Crush it, Game-changer, Guru, Ninja, Hack, Pivot, Synergy

TONE BY CONTEXT:
- Social Media: Conversational, witty, thought-provoking
- Client Presentations: Authoritative, polished, insight-led
- Case Studies: Results-focused, narrative-driven, specific
- Internal Comms: Transparent, collaborative, encouraging

REWRITING RULES:
1. Replace any BANNED words with the closest APPROVED alternative
2. Maintain the original message and intent — only change tone and word choice
3. Keep the same approximate length (within +/- 20%)
4. Remove jargon, buzzwords, and empty hype language
5. Add specificity where the original is vague (e.g., "great results" → "measurable impact")
6. If the tone is too casual, make it professional but warm
7. If the tone is too corporate/stiff, make it confident and human

OUTPUT FORMAT:
1. ORIGINAL: [paste the input copy]
2. ISSUES FOUND: [list each tone/word violation]
3. REWRITTEN COPY: [your corrected version]
4. CHANGES MADE: [brief bullet list of what you changed and why]

If the copy is already on-brand, respond: "This copy is on-brand. No changes needed."
"""


# --- Demo: Testing the enforcer with off-brand copy ---
import os
from openai import OpenAI
from dotenv import load_dotenv

load_dotenv()

client = OpenAI(
    api_key=os.getenv("OPENAI_API_KEY"),
    base_url="https://api.groq.com/openai/v1",
)


def enforce_brand_tone(off_brand_copy: str) -> str:
    """Rewrites off-brand copy to match Apex Digital Agency guidelines."""
    response = client.chat.completions.create(
        model="llama-3.3-70b-versatile",
        temperature=0.4,
        messages=[
            {"role": "system", "content": BRAND_TONE_ENFORCER_PROMPT},
            {"role": "user", "content": f"Rewrite this copy to match our brand guidelines:\n\n{off_brand_copy}"},
        ],
    )
    return response.choices[0].message.content


if __name__ == "__main__":
    # Test with off-brand copy examples
    test_copies = [
        "We're total marketing ninjas who crush it every day! Our game-changing strategies will disrupt your industry and create insane synergy across all channels! Let's hack growth together! 🚀🔥",
        "Our company does some stuff with data and AI. It's pretty cool I guess. We help companies or whatever. Hit us up if you want.",
        "LEVERAGING OUR PROPRIETARY SYNERGISTIC METHODOLOGIES, WE FACILITATE THE OPTIMIZATION OF CROSS-FUNCTIONAL PARADIGM SHIFTS TO MAXIMIZE STAKEHOLDER VALUE PROPOSITIONS.",
    ]

    for i, copy in enumerate(test_copies, 1):
        print(f"\n{'='*60}")
        print(f"TEST {i}")
        print(f"{'='*60}")
        result = enforce_brand_tone(copy)
        print(result)
