requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
PyMuPDF>=1.23.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
streamlit>=1.30.0
